7-kyu/maximum_length_difference/KataTest.java
package maximum_length_difference;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class KataTest {
    @Test
    void test() {
        System.out.println("mxdiflg Fixed Tests");
        String[] s1 = new String[]{"hoqq", "bbllkw", "oox", "ejjuyyy", "plmiis", "xxxzgpsssa", "xxwwkktt", "znnnnfqknaz", "qqquuhii", "dvvvwz"};
        String[] s2 = new String[]{"cccooommaaqqoxii", "gggqaffhhh", "tttoowwwmmww"};
        assertEquals(13, Kata.mxdiflg(s1, s2));
    }

    @Test
    void test2() {
        System.out.println("mxdiflg Fixed Tests");
        String[] s1 = new String[]{"ejjjjmmtthh", "zxxuueeg", "aanlljrrrxx", "dqqqaaabbb", "oocccffuucccjjjkkkjyyyeehh"};
        String[] s2 = new String[]{"bbbaaayddqbbrrrv"};
        assertEquals(10, Kata.mxdiflg(s1, s2));
    }
}
